Add observable property support
If observable property set and true, allow the user to specify an
input and an action if the input matches. Involved refactoring event
subscription logic as is very similar.

Also fix bug where ??? was treated as an array and would print a
bullet list when a value could not be read.